ai_news_crawler/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
schedule>=1.2.0
python-dateutil>=2.8.2
pytz>=2023.3
flask>=3.0.0
